requirements.txt
asgiref==3.4.1
astroid==2.4.2
cffi==1.14.1
cryptography==3.1.1
Django==3.2.6
django-cors-headers==3.5.0
django-extensions==3.1.0
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-storages==1.11.1
djangorestframework==3.11.0
gunicorn==20.1.0
isort==5.5.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-http-client==3.3.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.26.4
WebOb==1.8.7
whitenoise==5.2.0
wrapt==1.12.1
